Complete Student Info Controller, starting meeting info controller

+++ src/main/java/com/project/schoolmanagment/entity/concretes/business/StudentInfo.java
@@ -23,7 +23,7 @@ public class StudentInfo {
 
     private Integer absentee;
 
-    private Integer midtermExam;
+    private Double midtermExam;
 
     private Double finalExam;
